Follow up to initial commit (#7)
This change addresses some feedback from the initial commit that was not addressed
prior to merge.  Although it is good practice to satisfy all feedback prior to
merging, this practice was not followed for the initial commit.  This patch reflects
the changes that were suggested.

For starters, a typo in README.md was fixed to clarify documentation.  Additionally,
there was a small issue with how the "Runtime Dependencies" section of README.md
rendered on github's website, which diverged from how things looked on a local
GoLand based environment.  I included this change as it seemed apropos to other
changes being made.

Two packages were split off in order to promote better code separation and
dis-aggregation of separate functionality.  The "ansiblegalaxy" package was
introduced, and a package comment was created to reflect the theme as a wrapper
client to the ansible-galaxy binary.  During this conversion, it was discovered
that there were public constants that really ought not to be exposed outside the
package, so I lowered the visibility of these constants by making them begin with
a lower case letter.  A "helm" package was also introduced, and a package comment
was added surrounding its theme;  conversion from helm to Ansible Playbook Role.

Lastly, some updates to the Makefile were made in order to support the new
package structure.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>