Initial Branch Node Conversion (#3)
Branch structures include the control keywords "if" and "for".  These are common between
Go Templating and Jinja2.  However, the syntax of these structures varies fairly
significantly.  For instance, the following Go Template snippet:

```
  {{ if conditional }}
  ...
  {{ end }}
```

Should become:

```
  {% if conditional %}
  ...
  {% endif %}
```

This problem compounds since branch structures of any variety may be infinitely nested.
This commit extends the default Golang text/template module to output Jinja2 compliant
code.  This is just the start;  clearly there are other differences that must be addressed
too including whether "if conditional" means "if conditional is true" or
"if property is defined".  This is ambiguous and overloaded in Go templates, but must be
explicitly stated with Jinja2.

Additionally, the Helm funcs.go was modified and forked in, since Helm does not properly
expose/export their parser engine functionality in a way that a Golang client can directly
invoke.

Both of these decisions are explained in depth in their corresponding module README.md.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>