Expose functionality to help solve the "if-ambiguity" conversion problem
Helm defines templates utilizing Go "text/template" templates.  In Go "text/template",
the "if" conditional branch is overloaded, and can mean one of the following:
* "if <booleanExpression>"
* "if <property>"              # if property is defined.

Thus, during translation from Helm -> Ansible Template, the determination must be made
by taking a look at the Helm Chart's values.  If the key refers to a boolean, then
we must translate using boolean encoding logic.  If a key refers to anything else,
then it is likely non-boolean, and should be translated as "if <property> is defined".

This commit exposes a heuristic to expose whether a path refers to a boolean or not.
This is a heuristic, since sometimes Helm chart writers comment out configuration that
is not used.  Future commits will use this library when converting if-conditionals
from Golang templating to Jinja2 templating.

Several tests were abstracted to test this functionality using somewhat realistic
examples derived from the nginx helm chart.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>