Solve if-ambiguity and boolean-composition problems
Resolves "if-ambiguity" and "boolean-composition" problems in Go
Template to Jinja2 Template conversion.

1) if-ambiguity:  When converting a Go Template branch conditional like:
"{{ if .Values.something }}"
the meaning is ambiguous and can mean "if something is true" or
"if something is defined" depending on context.  This overloaded
behavior is not present in Jinja2.  Thus, conversion must rely on an
heuristic to determine whether the "something" is likely a boolean, or
an extended YAML definition.  This change implements such an heuristic
on terminal CommandNodes.

2) boolean-composition:  The boolean-composition problem is really two
sub-problems.  First, Go templates expect boolean composition in
functional form, such as "and condition1 condition2".  Jinja2, on the
other hand, expects "condition1 and condition2".  Additionally, these
expressions are often heavily nested, meaning a more comprehensive
and dynamic solution is needed.  This change implements boolean-
composition conversion by editing the underlying "text/template"
Abstract Syntax Tree.  When "and", "or" and "eq" nodes are encountered,
elements 0 and 1 are swapped such that the first condition becomes the
first CommandNode.

This change includes a number of different tests located in the forked
"text/template/parser", which include some contrived examples for
converting if-conditionals.  In order to keep the tests realistic,
minimal helm charts were built as input.  Each test parses input
charts directory, converts the parse trees for each template, then
compares the String() representation with an expected value.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>