Convert template function syntax to piped Jinja2 syntax
Jinja2 does not allow the following:
1) non-piped invocation
2) lack of parentheses/commas

This patch takes the Abstract Syntax Tree formed by the Go template engine
and converts the output to comply with the Jinja2 spec.  Expressions that
are non-piped are converted to utilize piped syntax.  Additionally, parens
and commas are inserted where appropriate.

A test case was added in order to demonstrate this new functionality.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>