Rewrite the Go Template to Jinja2 Template Parser
Essentially, Helm Ansible Template Exporter is a partial cross compiler
which takes Go Template input and exports Jinja2 syntax.  This impl. is
focused on a Helm origination context, so there is additional context
information made available to the compiler.

Prior to this patch, the implementation had gotten messy.  This was in
part due to the fact that the original Abstract Syntax Tree (AST) had
been maintained from the Go Template engine parser.  That was fine for
a demo, but lacks long term vision since translation methods were
largely hacked into existing overloaded methods.  Overloaded methods
refers to the fact that several Nodes were used for overloaded
functionality.  Namely, the BranchNode in the Go "text/template"
package is used to represent "if", "range" and "with" constructs.  Each
of these constructs can be translated into Jinja2, but each requires a
different translation path.  Thus, this patch removes BranchNode and
abstracts the following:

* IfNode:  Represents an "if" statement.  Has the knowledge to output
  "if" in a Jinja2 compliant manner.
* RangeNode:  Represents a "range" statement.  Has the functionality to
  convert Go "range" into a matching Jinja2 "for" statement.
* WithNode:  Represents a "with" statement.  With conversion is not
  supported in this version, so an appropriate warning is emitted to
  the user.

Additionally, PipeNode (canonically referred to as "pipeline" in the
original code) was special cased for these each "if", "range" and
"with" context.  This was done as the common ancestral parent of
"IfNode", "RangeNode" and "WithNode" is the "ActionNode", and in order
to appropriately convey context down the tree, each PipeNode needed
some special cased functionality.

An "IfCommandNode" was abstracted since conditional Commands are
much different than all other CommandNodes.  The primary difference is
that IfCommandNode's require a heuristic to infer the appropriate
translation into Jinja2.  Since this information was purely unique to
"IfNode", it was modified to have its own implementation so as to not
pollute other areas of code.

This change may appear rather large.  In fact, a lot of the code is
purely copy and paste given that many of the nodes do implement similar
functionality.  However, the differences in translation mechanism
were large enough to modify the AST in order to more cleanly render
Jinja2 output.

Lastly, the ugly j2Context argument was removed from the
Node.writeTo(...) interface.  IsConditional is now obvious since
conditionals have their own AST Node implementation (IfNode).
IsFunc is propagated from the parser through the FieldNode to
the underlying CommandNode.  PipeNodeCount was made a simple optional
property of the CommandNode abstraction.  This is much cleaner than
passing breadcrumbs across interfaces.

These changes were tested with the existing tests.  Additionally, they
were spot checked against the "nginx" and "redis" charts for
correctness.  A bunch of pre-existing tests that weren't broken from
these changes were re-enabled.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>