Optionally convert Ansible keys to snake_case
Unfortunately, operator-sdk converts Custom Resource (CR) keys to
snake_case.  This is a convention utilized by the operator-sdk team to
quickly identify Ansible keys.  However, this severely complicates
conversion from Helm Charts, as Helm Charts utilize Go Template
Language, and the convention in Go Template Language is camelCase.

Since this is a *hard* requirement, if exported Ansible Roles are
going to be used with operator-sdk (that is the aim of this project),
then we must adhere to the operator-sdk convention.  This first
requires converting Ansible Keys in defaults/main.yaml to snake_case
(see convert.ConvertDefaultsToSnakeCase).  Next, we need to parse the
templates for conversion, replacing FieldNodes with an appropriate
snake_case representation.  Since IfCommandNode, which can contain
FieldNode(s), consults defaults/main.yaml, we need to propagate logic
to perform the lookup based on the snake_case key, as the camelCase
key will no longer resolve.  Lastly, TextNodes in the AST must be
inspected for usages of converted yaml keys.  Since the representation
of a TextNode is arbitrary, and is essentially a blob of text,
everything that looks like a previously converted key is blindly
replaced.  In most cases, this is sufficient.  We do not have the
knowledge to determine the context of a TextNode, so this heuristic is
an appropriate heuristic.

All of this functionality is wrapped in a CLI option called
"emitKeysSnakeCase", and that key is set to true by default.  In most
use cases, this conversion is necessary to operate in an operator-sdk
context properly.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>