Update README.md and hack/README.md
Update the README.md with appropriate information about the two tools
provided.

The updates largely break the two tools down into corresponding
README.md files (one for the exporter, and one for the operator
builder).  These files explain corresponding parts in depth.  The user
can choose whether or not to follow one or both README.md(s).

A "Known Test Environments" section was added to the main README.md to
inform users what has worked for developers in the past.

Lastly, some information surrounding the "--exportKeysSnakeCase" flag
was provided.  Context was given as to operator-sdk conventions, and
why this flag is useful (and defaults to true).

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>